+++ src/modules/smoke/dto/smoke-error-query.dto.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsEnum } from 'class-validator';
+import { SmokeErrorTypeEnum } from '../enums/smoke-error-type.enum';
+
+export class SmokeErrorQueryDto {
+	@ApiProperty({
+		description: 'Error type to emulate',
+		enum: SmokeErrorTypeEnum,
+		example: SmokeErrorTypeEnum.BAD_REQUEST,
+	})
+	@IsEnum(SmokeErrorTypeEnum)
+	type!: SmokeErrorTypeEnum;
+}
