+++ src/modules/smoke/dto/smoke-validate-response.dto.ts
@@ -1,0 +1,3 @@
+import { SmokeValidateRequestDto } from './smoke-validate-request.dto';
+
+export class SmokeValidateResponseDto extends SmokeValidateRequestDto {}
