+++ src/modules/smoke/enums/smoke-error-type.enum.ts
@@ -1,0 +1,8 @@
+export enum SmokeErrorTypeEnum {
+	BAD_REQUEST = 'bad_request',
+	UNAUTHORIZED = 'unauthorized',
+	FORBIDDEN = 'forbidden',
+	NOT_FOUND = 'not_found',
+	CONFLICT = 'conflict',
+	INTERNAL = 'internal',
+}
