+++ src/modules/smoke/enums/smoke-health-status.enum.ts
@@ -1,0 +1,4 @@
+export enum SmokeHealthStatusEnum {
+	OK = 'ok',
+	ERROR = 'error',
+}
