+++ src/modules/smoke/index.ts
@@ -1,0 +1,1 @@
+export * from './module/smoke.module';
