+++ src/modules/smoke/services/smoke.service.ts
@@ -7,19 +7,17 @@ import {
 	NotFoundException,
 	UnauthorizedException,
 } from '@nestjs/common';
-import {
-	SmokeErrorQueryDto,
-	SmokeErrorType,
-	SmokeHealthResponseDto,
-	SmokeHealthStatus,
-	SmokeValidateRequestDto,
-	SmokeValidateResponseDto,
-} from './smoke.dto';
+import { SmokeValidateRequestDto } from '../dto/smoke-validate-request.dto';
+import { SmokeErrorQueryDto } from '../dto/smoke-error-query.dto';
+import { SmokeHealthResponseDto } from '../dto/smoke-health-response.dto';
+import { SmokeValidateResponseDto } from '../dto/smoke-validate-response.dto';
+import { SmokeHealthStatusEnum } from '../enums/smoke-health-status.enum';
+import { SmokeErrorTypeEnum } from '../enums/smoke-error-type.enum';
 
 @Injectable()
 export class SmokeService {
 	health(): SmokeHealthResponseDto {
-		return { status: SmokeHealthStatus.OK };
+		return { status: SmokeHealthStatusEnum.OK };
 	}
 	validate(body: SmokeValidateRequestDto): SmokeValidateResponseDto {
 		return body;
@@ -27,17 +25,17 @@ export class SmokeService {
 	getError(query: SmokeErrorQueryDto): void {
 		const errorText = `Smoke: ${query.type}`;
 		switch (query.type) {
-			case SmokeErrorType.BAD_REQUEST:
+			case SmokeErrorTypeEnum.BAD_REQUEST:
 				throw new BadRequestException(errorText);
-			case SmokeErrorType.UNAUTHORIZED:
+			case SmokeErrorTypeEnum.UNAUTHORIZED:
 				throw new UnauthorizedException(errorText);
-			case SmokeErrorType.FORBIDDEN:
+			case SmokeErrorTypeEnum.FORBIDDEN:
 				throw new ForbiddenException(errorText);
-			case SmokeErrorType.NOT_FOUND:
+			case SmokeErrorTypeEnum.NOT_FOUND:
 				throw new NotFoundException(errorText);
-			case SmokeErrorType.CONFLICT:
+			case SmokeErrorTypeEnum.CONFLICT:
 				throw new ConflictException(errorText);
-			case SmokeErrorType.INTERNAL:
+			case SmokeErrorTypeEnum.INTERNAL:
 				throw new InternalServerErrorException(errorText);
 			default:
 				throw new InternalServerErrorException(errorText);
