--- src/modules/smoke/smoke.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SmokeController } from './smoke.conroller';
-
-@Module({
-	controllers: [SmokeController],
-})
-export class SmokeModule {}
